commands/start.py
from telegram import Update
from telegram.ext import ContextTypes

async def handle_start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    await update.message.reply_text("👋 Привет! Я бот для анализа крипторынка. Используй команды, чтобы получить сигналы и аналитику.")